requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-openai
pypdf
python-docx
beautifulsoup4
langchain-community
faiss-cpu
gradio
python-dotenv
langchain-google-genai
pytest
